terrain map miniature

+++ CityGame/Data/DTO/GameObjectDTO.cs
@@ -4,7 +4,7 @@
     {
         public GroupDTO? Group { get; set; }
         public PositionDTO positionDTO { get; set; }
-        public int animationFrame { get; set; } = 1;
+        public byte animationFrame { get; set; } = 1;
 
         public byte level = 0;
